Split prometheus configuration
We're going to add some secrets to this file, and if we keep this as a
single file it's going to be very hard to manage since we would need to
move _all_ the configuration to the place where the secret is injected.

This change splits the prometheus configuration into separate files that
are stitched together when the container is started. That will allow the
single configuration section that has the secret to be managed
independently.